requirements.txt
selenium==4.17.2
beautifulsoup4==4.12.2
python-dotenv==1.0.1
Flask==2.3.3
Flask-HTTPAuth==4.8.0
Flask-RESTful==0.3.10
marshmallow==3.20.2
gunicorn==21.2.0